% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graticule_tiles.R
\name{tri_graticule}
\alias{tri_graticule}
\alias{quad_graticule}
\title{Graticule tiles}
\usage{
tri_graticule(n = 12, xlim = c(-180, 180), ylim = c(-90, 90),
  hull = FALSE, crs = "+proj=geocent +datum=WGS84")

quad_graticule(n = 12, xlim = c(-180, 180), ylim = c(-90, 90),
  hull = FALSE, crs = "+proj=geocent +datum=WGS84")
}
\arguments{
\item{n}{number of tile faces in x and y directions (give 2-elements for independent x, y)}
}
\value{
mesh3d object
}
\description{
Triangles, 2 per tile-face.
}
\details{
We use the termniology longitude (phi, azimuth) and latitude (theta, elevation) in the range
-180, 180 degrees. Longitude range is defined by xlim, latitude range by ylim.

We use the Geocentric XYZ projection by default, on WGS84. For spherical forms use
'crs = "+proj=sphere +a=1' for a sphere of radius 1.
}
\examples{
library(rgl)
x <- tri_graticule(n = c(12, 9))
clear3d()
shade3d(x)
aspect3d(1, 1, 1)
rglwidget()

library(rgl)
qx <- quad_graticule(n = c(12, 9))
clear3d()
shade3d(qx)
aspect3d(1, 1, 1)
rglwidget()
}
